requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
APScheduler==3.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
clarabel==0.11.1
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
curl_cffi==0.13.0
cvxpy==1.7.2
cycler==0.12.1
Cython==3.1.3
databricks-sdk==0.64.0
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
dnspython==2.7.0
docker==7.1.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.2
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
fonttools==4.59.2
fqdn==1.5.1
frozendict==2.4.6
futu_api==9.4.5408
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
gunicorn==23.0.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.9
lark==1.2.2
lightgbm==4.6.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mcp==1.13.1
mistune==3.1.3
mlflow==3.3.2
mlflow-skinny==3.3.2
mlflow-tracing==3.3.2
multitasking==0.0.12
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
osqp==1.0.4
packaging==25.0
pandas==2.3.2
pandocfilters==1.5.1
parso==0.8.5
peewee==3.18.2
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pymongo==4.14.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
pyqlib @ file:///home/kevin/Quanter/qlib
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
python-redis-lock==4.0.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.16.1
scs==3.2.8
Send2Trash==1.8.3
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlparse==0.5.3
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
TA-Lib==0.6.6
termcolor==3.1.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
yfinance==0.2.65
zipp==3.23.0
